Refactor media classes
- move media-player related classes to media_player.py
- create msg_definitions.py for internal WS msg data definitions
- Replace MediaType with MediaContentType
- Use StrEnum for all media-player enums
